+++ pkg/bk-monitor-worker/metrics/metadata.go
@@ -132,7 +132,6 @@ func MysqlCount(tableName, operation string, count float64) {
 	metric.Add(count)
 }
 
-
 func init() {
 	// register the metrics
 	Registry.MustRegister(
