+++ pkg/bk-monitor-worker/script/diff/cmd/root.go
