+++ pkg/bk-monitor-worker/script/diff/diff.yaml
